pyproject.toml
[project]
name = "courseproject2"
version = "0.1.0"
description = ""
authors = [
    {name = "Alexei",email = "streiker.saik@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pip (>=25.0.1,<26.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-cov = "^6.1.1"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.2.0"
black = "^25.1.0"
isort = "^6.0.1"
mypy = "^1.15.0"
types-requests = "^2.32.0.20250328"


[tool.black]
# Максимальная длина строки
line-length = 119
# Файлы, которые не нужно форматировать
exclude = '''
(
  /(
      \.git
  )/
)
'''

[tool.isort]
# максимальная длина строки
line_length = 119

[tool.mypy]
disallow_untyped_defs = true
warn_return_any = true
exclude = 'venv'